Fix: Comprehensive AI trump following behavior improvements
- Add extensive Issue #193 trump exhaustion bug tests covering trump pairs and tractors
- Add comprehensive Issue #176 trump pair disposal tests with conservation hierarchy
- Consolidate standalone issue176 test file into main trump following behavior test suite
- Import GamePhase enum for proper test setup
- Ensure AI properly exhausts all trump cards before non-trump when following trump leads
- Verify AI uses weakest trump pairs first according to conservation hierarchy
- Test trump rank pairs, trump suit pairs, and joker pairs in various scenarios
